Split up `API.User` into `API.User.*` for the test groups, plus `API.User.Prelude`.
