+++ services/brig/brig.cabal
@@ -322,7 +322,16 @@ executable brig-integration
         API.Team.Util
         API.TURN
         API.User
+        API.User.Prelude
+        API.User.Account
         API.User.Auth
+        API.User.Client
+        API.User.Connection
+        API.User.Handles
+        API.User.Invitation
+        API.User.Onboarding
+        API.User.PasswordReset
+        API.User.Property
         Util
 
     build-depends:
@@ -382,4 +391,3 @@ executable brig-integration
       , warp-tls              >= 3.2
       , zauth
       , yaml                  >= 0.8.22
-
